requirements.txt
Django==1.9.12
pymongo==3.5.0
psycopg2==2.7.3
djangorestframework==3.6.4
django-rest-framework-mongoengine==3.3.1
raven==6.1.0
cloudinary==1.8.0